+++ src/models/api/responses/collection_nft_trending.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
@@ -5,5 +6,5 @@ pub struct CollectionNftTrending {
     pub nft_image_url: Option<String>,
     pub nft_name: Option<String>,
     pub tx_frequency: Option<i64>,
-    pub last_price: Option<i64>,
+    pub last_price: Option<BigDecimal>,
 }
