+++ src/models/api/responses/data_point.rs
@@ -1,8 +1,9 @@
+use bigdecimal::BigDecimal;
 use chrono::{DateTime, Utc};
 use serde::Serialize;
 
 #[derive(Debug, Serialize)]
 pub struct DataPoint {
     pub x: Option<DateTime<Utc>>,
-    pub y: Option<i64>,
+    pub y: Option<BigDecimal>,
 }
