+++ src/models/api/responses/nft_listing.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
 pub struct NftListing {
-    pub price: Option<i64>,
+    pub price: Option<BigDecimal>,
     pub usd_price: Option<BigDecimal>,
     pub market_name: Option<String>,
     pub market_contract_id: Option<String>,
